classic to dark - last test
looking like change worked when changed the default in the scss file in css directory. Last change to make sure. This change shouldn't change the file from classic light to dark. If it doesn't I can safely delete all references to skin in this file and only change the skin by changing the default skin setting in the scss file in the css dir.